+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Road signs changelog
 
+
+## NEXT
+
+* Not published yet
+
+### Changes
+
+* This version is based on support library 28.0.0.
+* This version is based on Kotlin 1.3.61.
+
+
 ## [v.1.0.0](https://github.com/Umweltzone/roadsigns/releases/tag/v.1.0.0)
 
 * Published: 2019-08-30